changed some cursors to have additional flags, such as when writing the generated keys in the K2 and K3 building, the pages I write to now have the NO_FAULT flag. Not sure if this actually helps..
More importantly, I also avoid opening and closing cursors between searches. This may also speed things up a little bit. However, the uncompression times are slow for these bulk loading, so in this commit there is no compression in the K2 index, so that we can build the K3 index.