Include GPL license. Remove LZ4 code.
The GPL license at this point seems like the wise choice.

LZ4 was an experiment to see how well the handmade compression scheme worked. LZ4 does not perform was well, showing that our compression is better. In LZ4's defence, it is not a suitable compression scheme for this task, nor is any other compression scheme suitable, hence the need for doing compression our way.